rhel support & some fixes
