+++ pkg/configurer/enterpriselinux/el.go
@@ -1,0 +1,19 @@
+package enterpriselinux
+
+import (
+	"github.com/Mirantis/mcc/pkg/configurer"
+)
+
+// Configurer is the EL family specific implementation of a host configurer
+type Configurer struct {
+	configurer.LinuxConfigurer
+}
+
+// InstallBasePackages install all the needed base packages on the host
+func (c *Configurer) InstallBasePackages() error {
+	err := c.FixContainerizedHost()
+	if err != nil {
+		return err
+	}
+	return c.Host.Exec("sudo yum install -y curl")
+}
